app1/requirements.txt
fastapi
boto3
botocore
boto
yaml-1.3
ssl
uvicorn
fastapi.responses
fastapi.templating
jinja2
python-multipart

